1.4.3 (#1623)
* 🎵 NowPlayingCover not removed when fading between songs
- Removing duplicate artists from NowPlaying file
- Audio metadata dynamic value now remains when audio is paused
- Linting files

* 📺 Updated YT player package

* On https errors, show the url that errored.

* ✔️ Fixed custom group names reset when changing group of child slide
- Fixed add slide group with click not working
- Fixed bullet list size not resizing to font size, this also fixes very small font size in stage
- Small tweaks

* ✨ MediaShout mdb import
- Fixed MediaShout import not working
- Zooming in editors will keep slide more centered

* Properly awaiting chums tasks so we don't get multiple auth requests.

* 📂 Auto backup no longer backups shows content to save space
- Dropped templates added to correct position when creating slide
- Context menu enabled item checkmark
- Fixed creating a slide from template breaking dragging templates
- Fixed scripture sometimes behind from RemoteShow
- Fixed a scripture load issue
- Fixed dropping template on slide working when the show is locked
- Fixed project import location
- Tweaked rename popup
- Small tweaks

* ✨ Auto break long lines from template
- Variable tags
- Duplicate stage items
- Fixed cached cleared slide text next not showing up properly in stage displays
- Fixed playlist stop icon not clearing when clicked
- Fixed web search Genius lyrics getting song info
- Fixed context delete not working for all stage items
- Fixed API project item index not selected properly

* 🎨 Gradient color picker (#1613)

* ⚙️ More organized settings
- "Files" settings tab
- Removed "Groups" & "Cloud" tabs
- Open app data folder
- Moved manage metadata
- Setting gradient background transition
- Auto project name preview
- Autosave/backup time left preview
- Fixed freeze when condition timer is not defined
- Fixed large bullet points in lyrics view
- Name of stage items button "Slide text" changed when next slide text will be added
- Open quick search from menu bar
- Gradient works for stage items as well

* Logging failed requests

* Converted chums startup to a non-async method

* ✨ Enhancements
- Fixed list menu value not properly closing when reset
- Updated request error log
- String value to edit timer time works better
- Fixed some dynamic values not working in stage output
- Fixed conditions not accounting for the toggled state of a text variable
- Empty Chums project shows should be replaced by existing ones
- Fixed locked overlays not persisting if clear all before close
- Version update

---------

+++ src/electron/data/import.ts
@@ -13,8 +13,9 @@ import { dataFolderNames, doesPathExist, getDataFolder, getExtension, makeDir, r
 import { detectFileType } from "./bibleDetecter"
 import { filePathHashCode } from "./thumbnails"
 import { decompress, isZip } from "./zip"
+import MDBReader from "mdb-reader"
 
-type FileData = { content: Buffer | string; name?: string; extension?: string }
+type FileData = { content: Buffer | string | object; name?: string; extension?: string }
 
 const specialImports = {
     powerpoint: async (files: string[]) => {
@@ -68,6 +69,26 @@ const specialImports = {
 
         return data
     },
+    mdb: async (files: string[]) => {
+        const data: FileData[] = []
+
+        await Promise.all(files.map(async (filePath) => await mdbToFile(filePath)))
+
+        async function mdbToFile(filePath: string) {
+            const buffer = await readFileBufferAsync(filePath)
+            const reader = new MDBReader(buffer)
+            const tableNames = reader.getTableNames()
+
+            const dbData: any = {}
+            for (const tableName of tableNames) {
+                dbData[tableName] = reader.getTable(tableName).getData()
+            }
+
+            data.push({ content: dbData })
+        }
+
+        return data
+    },
 }
 
 // protobufjs (.pro) import can't handle close to 1000 files at once
@@ -82,6 +103,9 @@ export async function importShow(id: string, files: string[] | null, importSetti
     const sqliteFile = id === "openlp" && files.find((a) => a.endsWith(".sqlite"))
     if (sqliteFile) files = files.filter((a) => a.endsWith(".sqlite"))
     if (id === "easyworship" || id === "softprojector" || sqliteFile) importId = "sqlite"
+    const mdbFile = id === "mediashout" && files.find((a) => a.endsWith(".mdb"))
+    if (mdbFile) files = files.filter((a) => a.endsWith(".mdb"))
+    if (mdbFile) importId = "mdb"
 
     if (id === "freeshow_project") {
         await importProject(files, importSettings.path)
@@ -207,7 +231,7 @@ async function importProject(files: string[], dataPath: string) {
             if (!file) return
             replacedMedia[filePath] = newMediaPath
 
-            if (doesPathExist(newMediaPath)) return
+            if (doesPathExist(newMediaPath) || typeof file === "object") return
             writeFile(newMediaPath, file)
         })
 
@@ -246,6 +270,6 @@ async function decodeProto(filePath: string, fileContent: Buffer | null = null) 
 }
 
 async function checkSpecial(file: FileData) {
-    if (file.extension === "pro" && typeof file.content !== "string") return await decodeProto("", file.content)
+    if (file.extension === "pro" && Buffer.isBuffer(file.content)) return await decodeProto("", file.content)
     return
 }
